Refactor video_encoding.py by 1) rename to video_generation.py 2) rename class to VideoGeneration 3) delete unused functions 4) reformat code
